+++ memo-front/src/pages/article/article-list/article-list.html
@@ -1,0 +1,22 @@
+<!--
+  Generated template for the ArticleListPage page.
+
+  See http://ionicframework.com/docs/components/#navigation for more info on
+  Ionic pages and navigation.
+-->
+<ion-header>
+
+  <ion-navbar>
+    <ion-title>article-list</ion-title>
+  </ion-navbar>
+
+</ion-header>
+
+
+<ion-content padding>
+  <ion-list inset>
+    <button ion-item *ngFor="let article of articles" (click)="details(article.id)">
+      {{ article.title }}
+    </button>
+  </ion-list>
+</ion-content>
