+++ memo-front/src/pages/article/article-list/article-list.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { IonicPageModule } from 'ionic-angular';
+import { ArticleListPage } from './article-list';
+
+@NgModule({
+  declarations: [
+    ArticleListPage,
+  ],
+  imports: [
+    IonicPageModule.forChild(ArticleListPage),
+  ],
+})
+export class ArticleListPageModule {}
