+++ memo-front/src/pages/article/article-list/article-list.scss
@@ -1,0 +1,3 @@
+page-article-list {
+
+}
