--- memo-front/src/pages/article/article-listing/article-listing.page.html
@@ -1,1 +1,0 @@
-test
