+++ memo-front/src/pages/tabs/tabs.ts
@@ -3,7 +3,7 @@ import { Component } from '@angular/core';
 import { AboutPage } from '../about/about';
 import { ContactPage } from '../contact/contact';
 import { HomePage } from '../home/home';
-import { ArticleListingPage } from "../article/article-listing/article-listing.page";
+import { ArticleListPage } from "../article/article-list/article-list";
 
 @Component({
   templateUrl: 'tabs.html'
@@ -12,7 +12,7 @@ export class TabsPage {
 
   tab1Root = HomePage;
   tab2Root = AboutPage;
-  tab3Root = ArticleListingPage;
+  tab3Root = ArticleListPage;
   tab4Root = ContactPage;
 
   constructor() {
